# simplify: memory optimization