--- a/testing/dynamic_geoms_test.go
+++ b/testing/dynamic_geoms_test.go
@@ -1,6 +1,7 @@
 package testing
 
 import (
+	"math"
 	"testing"
 
 	"github.com/go-spatial/geom"
@@ -37,6 +38,38 @@ func TestBoxPolygon(t *testing.T) {
 		"dim -1" : {
 			dim: -1.0,
 			res: geom.Polygon{{{0, 0}, {-1.0, 0}, {-1.0, -1.0}, {0, -1.0}}},
+		},
+	}
+
+	for k, v := range tcases {
+		t.Run(k, fn(v))
+	}
+}
+
+func TestSinLineString(t *testing.T) {
+	type tcase struct {
+		amp, start, end float64
+		points int
+		out geom.LineString
+	}
+
+	fn := func (tc tcase) func(t *testing.T) {
+		return func(t *testing.T) {
+			out := SinLineString(tc.amp, tc.start, tc.end, tc.points)
+
+			if !cmp.GeometryEqual(out, tc.out) {
+				t.Errorf("expected %v got %v", tc.out, out)
+			}
+		}
+	}
+
+	tcases := map[string]tcase{
+		"amp 10" : {
+			amp: 10,
+			start: 0,
+			end: math.Pi * 2,
+			points: 5,
+			out: geom.LineString{{0, 0}, {math.Pi/2, 10}, {math.Pi, 0}, {3*math.Pi/2, -10}, {2*math.Pi, 0}},
 		},
 	}